refactor: ProfilePage fetching
